index config code is now in a separate file
